Rename OTLP exporter module to `otlp_grpc` and update provider to use the new module
