+++ src/exporters/mod.rs
@@ -13,7 +13,7 @@
 mod selectors;
 
 #[cfg(feature = "otlp")]
-pub mod otlp;
+pub mod otlp_grpc;
 
 #[cfg(feature = "stdout")]
 pub mod stdout;
